Patch Simbody/src/CablePath.cpp: Fix code
Looks like a problem of too many implicit conversions. The rules have probably
changed in recent standards. But the issue is Clang 15 on macOS which is quite
old so maybe it's because build2 enables the latest C++ standard (on an old
compiler!)  while upstream probably uses an older standard (better supported).

TODO Look into this later. Either way this is a temporary patch just to get CI
working.